// backend/src/domain/interface/users.interface.ts
export interface UserInterface {
    id?: number
    _id?: string
    email: string
    phone_no: string
    password: string
    name?: string
    no_hp?: string
    tenants?: string
    paket?: string
    passwordConfirmation?: string
    otp?: string
    is_active?: boolean
    paket_id?: number
    token?: string
    database?: string
    host?: string
    has_paid?: boolean
    hostname?: string
    sub_domain?: string
    token_verify?: string
    schemaname?: string
    template_id?: number
    is_admin?: boolean
    last_tenant?: number
    is_angpao_enabel?: boolean
    createdAt?: Date;
    updatedAt?: Date;
  }